lab2: fixxx iunt

+++ tasks/omp/kokin_marking_component_bin_image/func_tests/main.cpp
@@ -34,8 +34,8 @@ TEST(kokin_marking_component_bin_image, test_10_10) {
   int wh = 10;
   std::vector<int> size = {ht, wh};
   std::vector<uint8_t> in(ht * wh, 0);
-  std::vector<uint32_t> out(ht * wh, 0);
-  std::vector<uint32_t> cmpt(ht * wh, 1);
+  std::vector<uint8_t> out(ht * wh, 0);
+  std::vector<uint8_t> cmpt(ht * wh, 1);
 
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t *>(size.data()));
@@ -58,8 +58,8 @@ TEST(kokin_marking_component_bin_image, test_3_6) {
   int wh = 6;
   std::vector<int> size = {ht, wh};
   std::vector<uint8_t> in = {1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0};
-  std::vector<uint32_t> out(ht * wh, 0);
-  std::vector<uint32_t> cmpt = {0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 2, 0, 0, 0, 1, 1};
+  std::vector<uint8_t> out(ht * wh, 0);
+  std::vector<uint8_t> cmpt = {0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 2, 0, 0, 0, 1, 1};
 
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t *>(size.data()));
@@ -83,8 +83,8 @@ TEST(kokin_marking_component_bin_image, test_4) {
   std::vector<int> size = {ht, wh};
   std::vector<uint8_t> in = {0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1,
                              1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1};
-  std::vector<uint32_t> out(ht * wh, 0);
-  std::vector<uint32_t> cmpt = {1, 1, 0, 0, 2, 2, 2, 0, 0, 0, 1, 0, 0, 0, 0, 0, 2, 0, 0, 0, 1, 1, 0, 0, 0,
+  std::vector<uint8_t> out(ht * wh, 0);
+  std::vector<uint8_t> cmpt = {1, 1, 0, 0, 2, 2, 2, 0, 0, 0, 1, 0, 0, 0, 0, 0, 2, 0, 0, 0, 1, 1, 0, 0, 0,
                                 0, 0, 0, 0, 0, 0, 1, 0, 0, 3, 0, 3, 0, 0, 0, 0, 0, 0, 0, 3, 3, 3, 0, 0, 0};
 
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
@@ -108,8 +108,8 @@ TEST(kokin_marking_component_bin_image, test_5_5_arr) {
   int wh = 5;
   std::vector<int> size = {ht, wh};
   std::vector<uint8_t> in = {0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1};
-  std::vector<uint32_t> out(ht * wh, 0);
-  std::vector<uint32_t> cmpt = {1, 1, 0, 0, 2, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 3, 0, 0, 0, 3, 3, 0, 0, 0};
+  std::vector<uint8_t> out(ht * wh, 0);
+  std::vector<uint8_t> cmpt = {1, 1, 0, 0, 2, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 3, 0, 0, 0, 3, 3, 0, 0, 0};
 
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t *>(size.data()));
